requirements.txt
rich
psutil
pypresence
discord.py-self
colorama
inquirer
questionary
requests
art